fix warning (int/gsize mismatch with g_file_get_contents)
(Old svn revision: 17335)

+++ xfce4-about/info.c
@@ -317,7 +317,7 @@ add_page (GtkNotebook * notebook, const gchar * name, const gchar * filename, gb
   gchar *path;
   gchar *hfilename;
   gchar *buf;
-  int n;
+  gsize n;
 
   err = NULL;
